Move test it can calculate total calories to recipe.
