+++ test/cook_book_test.rb
@@ -29,25 +29,5 @@ class CookBookTest < Minitest::Test
     assert_equal [recipe1, recipe2], cookbook.recipes
   end
 
-  def test_it_can_calculate_total_calories
-    pantry = Pantry.new
-    cookbook = CookBook.new
-    ingredient1 = Ingredient.new({name: "Cheese", unit: "C", calories: 100})
-    ingredient2 = Ingredient.new({name: "Macaroni", unit: "oz", calories: 30})
-    recipe1 = Recipe.new("Mac and Cheese")
-
-    recipe1.add_ingredient(ingredient1, 2)
-    recipe1.add_ingredient(ingredient2, 8)
-
-    ingredient3 = Ingredient.new({name: "Ground Beef", unit: "oz", calories: 100})
-    ingredient4 = Ingredient.new({name: "Bun", unit: "g", calories: 75})
-    recipe2 = Recipe.new("Cheese Burger")
-
-    recipe2.add_ingredient(ingredient1, 2)
-    recipe2.add_ingredient(ingredient3, 4)
-    recipe2.add_ingredient(ingredient4, 1)
-
-    assert_equal 440, recipe1.total_calories
-    assert_equal 675, recipe2.total_calories
-  end
+  
 end
